backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pdfplumber==0.10.3
python-docx==1.1.0
google-generativeai>=0.8.0
python-multipart==0.0.6
requests==2.32.3
python-dotenv==1.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.19.0

